+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## master (unreleased)
 
+## 2.0.1 (2019-06-08)
+
+### Changes
+
+* [#68](https://github.com/rubocop-hq/rubocop-rails/pull/68): Relax rack dependency for Rails 4. ([@buehmann][])
+
 ## 2.0.0 (2019-05-22)
 
 ### New features
@@ -11,3 +17,4 @@
 
 [@koic]: https://github.com/koic
 [@andyw8]: https://github.com/andyw8
+[@buehmann]: https://github.com/buehmann